pyproject.toml
[project]
name = "timage-public"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.26.0",
    "celery[redis]>=5.6.2",
    "cryptg>=0.5.2",
    "django>=6.0.3",
    "django-celery-beat>=2.9.0",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "django-minio-backend>=4.5.0",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "gunicorn>=25.1.0",
    "imagehash>=4.3.2",
    "markdown>=3.10.2",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "psycopg[binary,pool]>=3.3.3",
    "pyngrok>=7.5.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "sentry-sdk[django]>=2.54.0",
    "telethon>=1.42.0",
    "uuid6>=2025.0.1",
    "uvicorn>=0.42.0",
    "whitenoise>=6.12.0",
]
